+++ src/main/java/net/kaczmarzyk/spring/data/jpa/domain/Join.java
@@ -29,6 +29,8 @@ import jakarta.persistence.criteria.Root;
 import static net.kaczmarzyk.spring.data.jpa.utils.JoinPathUtils.pathToJoinContainsAlias;
 import static net.kaczmarzyk.spring.data.jpa.utils.JoinPathUtils.pathToJoinSplittedByDot;
 
+import java.util.function.Function;
+
 /**
  * @author Tomasz Kaczmarzyk
  * @author Jakub Radlica
@@ -58,7 +60,7 @@ public class Join<T> implements Specification<T>, Fake {
 
 		if (!pathToJoinContainsAlias(pathToJoinOn)) {
                         if(!queryContext.existsJoin(alias, root)) {
-                            queryContext.putLazyVal(alias, (r) -> r.join(pathToJoinOn, joinType));
+                        	putValToQueryContext(alias, root, (r) -> r.join(pathToJoinOn, joinType));
                         }
 		} else {
 			String[] pathToJoinOnSplittedByDot = pathToJoinSplittedByDot(pathToJoinOn);
@@ -73,8 +75,9 @@ public class Join<T> implements Specification<T>, Fake {
 			}
 
 			String extractedPathToJoin = pathToJoinOnSplittedByDot[1];
-                queryContext.putLazyVal(
+				putValToQueryContext(
                         alias,
+                        root,
                         (r) -> {
                         	jakarta.persistence.criteria.Join<?, ?> evaluated = queryContext.getEvaluated(extractedAlias, root);
                         	return evaluated.join(extractedPathToJoin, joinType);
@@ -82,6 +85,19 @@ public class Join<T> implements Specification<T>, Fake {
                 );
 		}
 		return null;
+	}
+
+	private void putValToQueryContext(String alias, Root<T> root, Function<Root<?>, jakarta.persistence.criteria.Join<?, ?>> lazyVal) {
+		// generally we want to evaluate join lazily
+		// because most typical scenario tends to be a LEFT join with distinct = true
+		// and in such scenario if there is no filtering on the joined part (e.g. no related http param was sent)
+		// then we can optimize behaviour by not joining at all
+		queryContext.putLazyVal(alias, lazyVal);
+		// but inner joins or non-distinct queries must have them evaluated eagerly
+		// because they affect query result even when there is no filtering applied
+		if (!distinctQuery || joinType == JoinType.INNER) {
+			queryContext.getEvaluated(alias, root);
+		}
 	}
 
 	@Override