+++ src/main/java/net/kaczmarzyk/spring/data/jpa/domain/True.java
@@ -20,7 +20,7 @@ import net.kaczmarzyk.spring.data.jpa.utils.QueryContext;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
-import jakarta.persistence.criteria.Root;;
+import jakarta.persistence.criteria.Root;
 import java.util.Arrays;
 import java.util.Objects;