+++ src/test/java/net/kaczmarzyk/spring/data/jpa/OrderNote.java
@@ -17,6 +17,7 @@ package net.kaczmarzyk.spring.data.jpa;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToOne;
 
@@ -24,7 +25,7 @@ import jakarta.persistence.OneToOne;
 public class OrderNote {
 
 	@Id
-	@GeneratedValue
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	private String title;