+++ src/test/java/net/kaczmarzyk/spring/data/jpa/Person.java
@@ -17,13 +17,14 @@ package net.kaczmarzyk.spring.data.jpa;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 
 @Entity
 public class Person {
 
 	@Id
-	@GeneratedValue
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	
 	private String name;